MBS-4517 remove funktionale usage in :test-report
